Move the logic into `__main__.py` so the module can be executed with `python -m passphraseme`
